requirements.txt
numpy
sounddevice
scipy
speech_recognition
midiutil
python-dotenv
pytest